pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Alexander Maksimovich"]
packages = [
    { include = "gendiff" }
]


[tool.poetry.dependencies]
python = "^3.8"
argparse = "^1.4.0"
pytest-cov = "^2.12.1"
PyYAML = "^5.4.1"


[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
flake8 = "^3.9.2"
coverage = "^5.5"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"
